types.h: bump integer types to uint64_t

+++ src/book/types.h
@@ -9,12 +9,12 @@
 
 namespace liquibook { namespace book {
   // Types used in Liquibook
-  typedef uint32_t Price;
-  typedef uint32_t Quantity;
-  typedef uint32_t Cost;
-  typedef uint32_t FillId;
-  typedef uint32_t ChangeId;
-  typedef uint32_t OrderConditions;
+  typedef uint64_t Price;
+  typedef uint64_t Quantity;
+  typedef uint64_t Cost;
+  typedef uint64_t FillId;
+  typedef uint64_t ChangeId;
+  typedef uint64_t OrderConditions;
 
   enum OrderCondition {
     oc_no_conditions = 0,
@@ -28,7 +28,7 @@ namespace liquibook { namespace book {
   // Constants used in liquibook API
   const Price MARKET_ORDER_PRICE(0);
   const Price PRICE_UNCHANGED(0);
-  const int32_t SIZE_UNCHANGED(0);
+  const int64_t SIZE_UNCHANGED(0);
   }
 
 } } // namespace
